stock-analyzer-agent/pyproject.toml
[project]
name = "stock-analyzer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.7",
    "numexpr>=2.14.1",
    "python-dotenv>=1.2.1",
    "yfinance>=1.1.0",
]
